requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
fonttools==4.51.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.4.16
requests==2.31.0
six==1.16.0
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
vaderSentiment==3.3.2
Werkzeug==3.0.2
wordcloud==1.9.3
zipp==3.18.1
